Additional ReadMe Tasks/Comments
John and Abby have been working on isolating the resume opening bug for the last two weeks however, due to scheduling conflicts and issues regarding this set of npm install and firebase, we were slowed down. We are back on track to replicate the bug and work on fixing it. Ian has sent an example of a resume that is broken, on John machine it successfully opens however, Ian clarified that the user uploaded the resume on a Mac Book pro through Chrome (which is the set up that Abby has). Our next steps are the get the resume uploaded through Abby's computer and locate the bug that way.